feat: Organize appland client API by object type
