--- packages/client/src/getAppMap.ts
@@ -1,37 +1,0 @@
-import { IncomingMessage } from 'node:http';
-import { URL } from 'node:url';
-import { AppMap } from '@appland/models';
-import Settings from './settings';
-import reportJSON from './reportJson';
-import validateSettings from './validateSettings';
-import handleError from './handleError';
-import buildRequest from './buildRequest';
-
-/**
- * Loads AppMap data from UUID.
- */
-export default function getAppMap(uuid: string): Promise<AppMap> {
-  const call = (): Promise<IncomingMessage> => {
-    return new Promise((resolve, reject) => {
-      const request = buildRequest();
-      const getScenarioURL = new URL(
-        [Settings.baseURL, 'api/appmaps', uuid].join('/')
-      );
-      request
-        .requestFunction(
-          getScenarioURL,
-          {
-            headers: request.headers,
-          },
-          resolve
-        )
-        .on('error', reject)
-        .end();
-    });
-  };
-
-  return validateSettings()
-    .then(call)
-    .then(handleError)
-    .then((response) => reportJSON<AppMap>(response));
-}
