+++ packages/client/src/index.ts
@@ -1,5 +1,5 @@
-export { default as getAppMap } from './getAppMap';
-export { default as listAppMap } from './listAppMaps';
-export { default as listFindingStatus } from './listFindingStatus';
+export { default as App } from './app';
+export { default as AppMap } from './appMap';
+export { default as Mapset } from './mapset';
 export { default as Settings } from './settings';
 export { AppMapListItem, FindingStatusListItem } from './types';
